finished  administration

+++ Areas/Admin/Controllers/OrderItemController.cs
@@ -18,29 +18,6 @@ namespace web_shop_app.Areas.Admin.Controllers
             _context = context;
         }
 
-        // GET: Admin/OrderItem
-        public async Task<IActionResult> Index()
-        {
-            return View(await _context.OrderItems.ToListAsync());
-        }
-
-        // GET: Admin/OrderItem/Details/5
-        public async Task<IActionResult> Details(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var orderItem = await _context.OrderItems
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (orderItem == null)
-            {
-                return NotFound();
-            }
-
-            return View(orderItem);
-        }
 
         // GET: Admin/OrderItem/Create
         public IActionResult Create(int orderId)
@@ -76,11 +53,24 @@ namespace web_shop_app.Areas.Admin.Controllers
                     return NotFound();
                 }
 
-                orderItem.Total = orderItem.Quantity * productDetails.Price;
+                var order = await _context.Orders.FirstOrDefaultAsync(o => o.Id == orderItem.OrderId);
 
-                _context.Add(orderItem);
+                if (order == null)
+                {
+                    return NotFound();
+                }
+
+                orderItem.Total = orderItem.Quantity * productDetails.Price;
+                order.Total = order.Total + orderItem.Total;
+
+                // order.Total += orderItem.Total --- kraće----
+
+
+                await _context.AddAsync(orderItem);
+                 _context.Update(order);
+                
                 await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                return RedirectToAction("Details","Order", new { id =  orderItem.OrderId});
             }
             catch (Exception)
             {
@@ -90,52 +80,6 @@ namespace web_shop_app.Areas.Admin.Controllers
           
         }
 
-        // GET: Admin/OrderItem/Edit/5
-        public async Task<IActionResult> Edit(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var orderItem = await _context.OrderItems.FindAsync(id);
-            if (orderItem == null)
-            {
-                return NotFound();
-            }
-            return View(orderItem);
-        }
-
-        // POST: Admin/OrderItem/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,OrderId,ProductId,Quantity,Total")] OrderItem orderItem)
-        {
-            if (id != orderItem.Id)
-            {
-                return NotFound();
-            }
-
-            try
-            {
-                _context.Update(orderItem);
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!OrderItemExists(orderItem.Id))
-                {
-                    return NotFound();
-                }
-                else
-                {
-                    return View(orderItem);
-                }
-            }
-            return RedirectToAction(nameof(Index));
-        }
 
         // GET: Admin/OrderItem/Delete/5
         public async Task<IActionResult> Delete(int? id)
@@ -160,14 +104,31 @@ namespace web_shop_app.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var orderItem = await _context.OrderItems.FindAsync(id);
+             var orderItem = await _context.OrderItems.FindAsync(id);
+
+            if(orderItem == null)
+            {
+                return NotFound();
+            }
+
+            var order = await _context.Orders.FindAsync(orderItem.OrderId);
+
+            if (order == null)
+            {
+                return NotFound();
+            }
+
+            order.Total = order.Total - orderItem.Total;
+
             if (orderItem != null)
             {
                 _context.OrderItems.Remove(orderItem);
             }
 
+            _context.Update(order);
+
             await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            return RedirectToAction("Details", "Order", new { id = orderItem.OrderId });
         }
 
         private bool OrderItemExists(int id)
